Refactoring the swap creation logic into SwapStore
